requirements-prod.txt
# Production Requirements
# This file contains only production dependencies (no testing libraries)

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic[email]==2.5.0
email-validator==2.1.0
httpx==0.25.2
passlib[argon2]==1.7.4
argon2-cffi
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

